fix: split feature-flags into client-safe shared module — fixes Turbopack build
posthog-node (server-only) was being pulled into client bundles via
use-feature-flag.ts → feature-flags.ts import chain. Split shared types
and static fallbacks into feature-flags-shared.ts (client-safe).
